move error and warning messages above changes list in stop-loss
